requirements.txt
djangorestframework-simplejwt==5.2.2